Combine Vertical Remapping into a single object
