+++ tests/translate/__init__.py
@@ -63,8 +63,6 @@ from .translate_qsinit import TranslateQSInit
 from .translate_ray_fast import TranslateRay_Fast
 from .translate_remap_profile_2d import TranslateCS_Profile_2d, TranslateCS_Profile_2d_2
 from .translate_remapping import TranslateRemapping
-from .translate_remapping_part1 import TranslateRemapping_Part1
-from .translate_remapping_part2 import TranslateRemapping_Part2
 from .translate_riem_solver3 import TranslateRiem_Solver3
 from .translate_riem_solver_c import TranslateRiem_Solver_C
 from .translate_satadjust3d import TranslateSatAdjust3d
